perf(frontend): load the changelog drawer's content on demand
Split ChangelogPanel.tsx into ChangelogPanel.tsx (a shim) and
ChangelogPanelContent.tsx (the moved implementation, verbatim except for
two edits: the props interface and the default export are now exported so
the shim can import them).

changelog-data.ts is 432 KB raw / 131 KB gzipped across 102 releases,
about 19% of what a visitor downloads. LoginChoice.tsx renders this panel,
so before this split an unauthenticated visitor downloaded the project's
entire engineering diary before they could even log in.

The shim wraps the content in React.lazy()/Suspense so the data-heavy
module is only fetched once the drawer actually opens. The Suspense
boundary lives in the shim rather than in the package because
@ronl/pa-cockpit's PADashboardV2 renders this panel through the host
contract with no Suspense boundary of its own anywhere in either package;
a bare lazy() export would throw the moment the drawer opened on the PA
cockpit route, and adding <Suspense> inside the package would leak this
host's implementation choice into a contract that packages/pa-demo also
depends on without needing it. The shim's `if (!isOpen) return null`
early return is load-bearing, not stylistic: rendered unconditionally,
the lazy element starts fetching the chunk on every page load and the
split saves nothing.

Test file: repointed the ScopeBadge import to ChangelogPanelContent
(it has no consumer outside this file and its test) and added `await
screen.findByRole(...)` to the seven ChangelogPanel tests that assert on
content, per a diagnostic run.

That diagnostic surfaced a subtlety worth recording. The shim's
`lazy(() => import('./ChangelogPanelContent'))` call sits at module scope,
so its resolved-promise cache is shared by every test in the file. The
first test to render the drawer open triggers and resolves that import;
every test that runs afterward in the same file then renders the content
synchronously and would pass regardless of whether it awaits, because it
is riding the first test's cache warm-up rather than exercising its own
render path. A whole-file run therefore showed only 1 of 8 content tests
failing before this fix, not 7 — but running each of the other six alone
(so the cache starts cold) reproduced the same synchronous-assertion
failure in all of them. All seven now await the content explicitly, so
each holds on its own, in isolation, and under reordering rather than by
accident of which test happened to run first. A comment on the first
await records this so a future reader who sees six of them "pass without
awaiting" during some partial run doesn't conclude they were redundant
and delete them.

No behavioural change: all five call sites, the host contract, and
packages/pa-cockpit and packages/pa-demo are untouched. changelog-data.ts
is untouched; all 102 releases still ship, just no longer inside the
entry chunk. Frontend suite: 842 tests / 103 files, unchanged from
baseline — this task adds no tests, it only changes how seven existing
ones observe an async render.
